requirements/base.txt
arrow==0.7.0
beautifulsoup4==4.4.1
bs4==0.0.1
Django==1.10.1
django-import-export==0.5.0
django-haystack==2.5.0
django-rest-swagger==2.0.5
djangorestframework==3.4.1
djangorestframework-csv==1.4.1
elasticsearch==2.3.0
Markdown==2.6.6
psycopg2==2.6.1
python-dateutil==2.5.2
PyYAML==3.11
six==1.10.0
sqlparse==0.1.19
tablib==0.11.2
Unipath==1.1
urllib3==1.14
